rename: Decimal → Float (the float64 numeric type)
The float64 leaf of Number was misleadingly named `Decimal` (Exhibit L
of the WAT audit): it is binary IEEE-754, so `0.1 add 0.2` is not
`0.3`. Rename it to `Float` and free the name `Decimal` for a future
true base-10 type (none added here).

Scope:
- Kernel + library Go: type path `Scalar/Number/Decimal` →
  `Scalar/Number/Float`, short name `Decimal` → `Float`, and the
  identifiers `TDecimal`/`NewDecimal`/`AsDecimal`/`AsConcreteDecimal`/
  `DecPayload`/`formatDecimal`/`DepDecimal`/… → their `Float` forms.
- FixedID: the type keeps FixedID 9 (only its path string changes), so
  serialized Value IDs and the wire format are unchanged; the stability
  snapshot is updated to the new path.
- User-facing strings: `convert`/`make` "to decimal" → "to float",
  formatter known-type set, and `describe` help text.
- Spec: all eng/spec + lang/spec + lang/go/test TSV expectations.
- Docs: README/TUTORIAL/EXPLANATION/REFERENCE/HOWTO/FORMAL-SPEC type
  references (leaving base-10 "decimal point/notation" wording); the
  REFERENCE note now reads "`Float` is IEEE-754 float64" and records
  that `Decimal` is reserved. Kernel CLAUDE.md guidance updated.

`typeof 3.14` now returns `Float`; `Decimal` is undefined (reserved).
All modules build; eng/lang/test/calc/cmd suites and the doc/CLI
example harnesses pass. Comments preserving base-10 "decimal" notation
(lexer literal parsing, word-modifier arg counts) are intentionally
left unchanged. design/ historical notes are left as-is.

+++ lang/go/native/help/help.go
@@ -369,7 +369,7 @@ func writeSigs(b *strings.Builder, sigs []SigInfo) {
 // Standard example values by type leaf name.
 var exampleValues = map[string][]string{
 	"Integer": {"2", "3", "4", "5", "6", "7", "8", "9"},
-	"Decimal": {"2.25", "3.5", "4.75", "5.25", "6.5", "7.75", "8.25", "9.5"},
+	"Float": {"2.25", "3.5", "4.75", "5.25", "6.5", "7.75", "8.25", "9.5"},
 	"Number":  {"2", "3.5", "4", "5.25", "6", "7.5", "8", "9.25"},
 	"String":  {"'a'", "'b'", "'c'", "'d'", "'e'", "'f'"},
 	"Scalar":  {"'a'", "'b'", "'c'", "'d'", "'e'", "'f'"},
